requirements.txt
selenium>=4.0.0
colorama>=0.4.4
tqdm>=4.62.3
schedule>=1.1.0
python-dotenv>=0.19.2
requests>=2.26.0